requirements.txt
asgiref==3.8.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Django==5.1.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.0
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
SQLAlchemy==2.0.40
sqlparse==0.5.1
typing_extensions==4.13.2
tzdata==2024.2
uritemplate==4.1.1
Werkzeug==3.1.3
